ux_dashboard/requirements.txt
boto3==1.26.59
boto3-stubs-lite==1.26.64
Flask==2.2.2
psycopg2-binary==2.9.6
python-dotenv==0.21.1